Somewhat finished
I have made the computer single player version and multiplayer version. Input validation should be fully fixed and won't allow players to work around it. Want to teach the computer mode tricks to look for so it's practically impossible to win in single player. Might even give the option to allow computer to go first later on. Not much else otherwise.