Player 1 is too powerful
I've tried many variations to test if player 2 can win and I've found it pretty much impossible to do so if player 1 starts in the middle. I found counters to counters. Might sleep on this and try thinking about it again tomorrow, but I can't find an optimal way the computer can even tie this with my different counters I find.